Enabled Quest 3, minor update to quest 2
Quest 3 call functions for the quest was disabled before. I have enabled it.
I took out the extra brackets ( {} ) that I had separating the scenes. It was mostly just for me to be able to more easily see what I was doing and at what point in the code I was at. They might have caused us problems if they I did not take them out.