Visual studio
I got tired of the IDE I've been using (DEV C++). It wasn't telling me
anything worth my time on what was wrong. I created a visual studio
project and here it is. I haven't tested if it works in this folder or
not yet but it works just fine on my computer. I have completed most of
what I needed to for the classes and stuff.